+++ NEWS.md
@@ -20,6 +20,7 @@ Fixes:
 * Fixed references for missing-value implementation mockup in vignette.
 * Fixed a bug in `normalise` that resulted in relations having duplicate keys.
 * Fixed a bug in `normalise`, that resulted in schemas that didn't reproduce the given functional dependencies.
+* Fixed `dfd`'s data simplification step for POSIXct datetimes, in case where two times only differ by standard/daylight-savings time (e.g. 1:00:00 EST vs. 1:00:00 EDT on the same day).
 
 # autodb 1.0.0